move bucket sort to query_builder
